@@ -2,6 +2,5 @@ package com.meterial.drawer.activity;
 
 import com.yline.application.BaseApplication;
 
-public class IApplication extends BaseApplication
-{
+public class IApplication extends BaseApplication {
 }
